git-svn-id: http://googleappengine.googlecode.com/svn/trunk@277 80f5ef21-4148-0410-bacc-cfb02402ada8

+++ python/google/appengine/ext/datastore_admin/backup_handler.py
@@ -30,13 +30,19 @@ page.
 This module also contains actual mapper code for backing data over.
 """
 
+from __future__ import with_statement
 
+
+
+import cStringIO
 import datetime
 import itertools
 import logging
+import os
 import re
 import time
 import urllib
+import xml.dom.minidom
 
 from google.appengine.datastore import entity_pb
 from google.appengine.api import apiproxy_stub_map
@@ -47,6 +53,7 @@ from google.appengine.api import datastore
 from google.appengine.api import files
 from google.appengine.api import taskqueue
 from google.appengine.api import urlfetch
+from google.appengine.api.files import records
 from google.appengine.api.taskqueue import taskqueue_service_pb
 from google.appengine.ext import blobstore
 from google.appengine.ext import db
@@ -88,14 +95,12 @@ class ConfirmBackupHandler(webapp.RequestHandler):
     notreadonly_warning = capabilities.CapabilitySet(
         'datastore_v3', capabilities=['write']).is_enabled()
     blob_warning = bool(blobstore.BlobInfo.all().count(1))
-    app_id = handler.request.get('app_id')
     template_params = {
         'form_target': DoBackupHandler.SUFFIX,
         'kind_list': kinds,
         'remainder': remainder,
         'sizes_known': sizes_known,
         'size_total': size_total,
-        'app_id': app_id,
         'queues': None,
         'cancel_url': handler.request.get('cancel_url'),
         'has_namespace': has_namespace,
@@ -130,7 +135,6 @@ class ConfirmDeleteBackupHandler(webapp.RequestHandler):
           gs_warning |= backup.filesystem == files.GS_FILESYSTEM
     template_params = {
         'form_target': DoBackupDeleteHandler.SUFFIX,
-        'app_id': handler.request.get('app_id'),
         'cancel_url': handler.request.get('cancel_url'),
         'backups': backups,
         'xsrf_token': utils.CreateXsrfToken(XSRF_ACTION),
@@ -160,7 +164,6 @@ class ConfirmAbortBackupHandler(webapp.RequestHandler):
           backups.append(backup)
     template_params = {
         'form_target': DoBackupAbortHandler.SUFFIX,
-        'app_id': handler.request.get('app_id'),
         'cancel_url': handler.request.get('cancel_url'),
         'backups': backups,
         'xsrf_token': utils.CreateXsrfToken(XSRF_ACTION)
@@ -175,27 +178,87 @@ class ConfirmRestoreFromBackupHandler(webapp.RequestHandler):
   SUFFIX = 'confirm_restore_from_backup'
 
   @classmethod
+  def Render(cls, handler, default_backup_id=None,
+             default_delete_backup_after_restore=False):
+    """Rendering method that can be called by main.py.
+
+    Args:
+      handler: the webapp.RequestHandler invoking the method
+      default_backup_id: default value for handler.request
+      default_delete_backup_after_restore: default value for handler.request
+    """
+    backup_id = handler.request.get('backup_id', default_backup_id)
+    backup = db.get(backup_id) if backup_id else None
+    notreadonly_warning = capabilities.CapabilitySet(
+        'datastore_v3', capabilities=['write']).is_enabled()
+    original_app_warning = backup.original_app
+    if os.getenv('APPLICATION_ID') == original_app_warning:
+      original_app_warning = None
+    template_params = {
+        'form_target': DoBackupRestoreHandler.SUFFIX,
+        'queues': None,
+        'cancel_url': handler.request.get('cancel_url'),
+        'backup': backup,
+        'delete_backup_after_restore': handler.request.get(
+            'delete_backup_after_restore', default_delete_backup_after_restore),
+        'xsrf_token': utils.CreateXsrfToken(XSRF_ACTION),
+        'notreadonly_warning': notreadonly_warning,
+        'original_app_warning': original_app_warning
+    }
+    utils.RenderToResponse(handler, 'confirm_restore_from_backup.html',
+                           template_params)
+
+
+class ConfirmBackupImportHandler(webapp.RequestHandler):
+  """Handler to import backup information."""
+
+  SUFFIX = 'backup_information'
+
+  @classmethod
   def Render(cls, handler):
     """Rendering method that can be called by main.py.
 
     Args:
       handler: the webapp.RequestHandler invoking the method
     """
-    backup_id = handler.request.get('backup_id')
-    backup = db.get(backup_id) if backup_id else None
-    notreadonly_warning = capabilities.CapabilitySet(
-        'datastore_v3', capabilities=['write']).is_enabled()
-    app_id = handler.request.get('app_id')
+    gs_handle = handler.request.get('gs_handle')
+    error = None if gs_handle else 'Google Cloud Storage path is missing'
+    other_backup_info_files = []
+    selected_backup_info_file = None
+    backup_info_specified = False
+    if not error:
+      try:
+        gs_handle = gs_handle.rstrip()
+        bucket_name, prefix = parse_gs_handle(gs_handle)
+        validate_gs_bucket_name(bucket_name)
+        if not is_accessible_bucket_name(bucket_name):
+          raise BackupValidationException(
+              'Bucket "%s" is not accessible' % bucket_name)
+        if prefix.endswith('.backup_info'):
+          prefix = prefix[0:prefix.rfind('/')]
+          backup_info_specified = True
+        elif prefix and not prefix.endswith('/'):
+          prefix += '/'
+        for backup_info_file in list_bucket_files(bucket_name, prefix):
+          if backup_info_file.endswith('.backup_info'):
+            backup_info_file = '/gs/%s/%s' % (bucket_name, backup_info_file)
+
+            if backup_info_specified and backup_info_file == gs_handle:
+              selected_backup_info_file = backup_info_file
+            else:
+              other_backup_info_files.append(backup_info_file)
+      except Exception, ex:
+        error = 'Failed to read bucket: %s' % ex
     template_params = {
-        'form_target': DoBackupRestoreHandler.SUFFIX,
-        'app_id': app_id,
-        'queues': None,
+        'error': error,
+        'form_target': DoBackupImportHandler.SUFFIX,
         'cancel_url': handler.request.get('cancel_url'),
-        'backup': backup,
-        'xsrf_token': utils.CreateXsrfToken(XSRF_ACTION),
-        'notreadonly_warning': notreadonly_warning
+        'selected_backup_info_file': selected_backup_info_file,
+        'other_backup_info_files': other_backup_info_files,
+        'backup_info_specified': backup_info_specified,
+        'xsrf_token': utils.CreateXsrfToken(XSRF_ACTION)
     }
-    utils.RenderToResponse(handler, 'confirm_restore_from_backup.html',
+    utils.RenderToResponse(handler, 'confirm_backup_import.html',
                            template_params)
 
 
@@ -334,7 +397,8 @@ def _perform_backup(kinds,
 
     if not gs_bucket_name:
       raise BackupValidationException('Bucket name missing.')
-    bucket_name = gs_bucket_name.split('/', 1)[0]
+    bucket_name, path = parse_gs_handle(gs_bucket_name)
+    gs_bucket_name = ('%s/%s' % (bucket_name, path)).rstrip('/')
     validate_gs_bucket_name(bucket_name)
     if not is_accessible_bucket_name(bucket_name):
       raise BackupValidationException(
@@ -533,15 +597,16 @@ def delete_backup_files(filesystem, backup_files):
         blobstore_api.delete(blob_keys)
 
 
-def delete_backup_info(backup_info):
+def delete_backup_info(backup_info, delete_files=True):
   """Deletes a backup including its associated files and other metadata."""
   if backup_info.blob_files:
     delete_backup_files(backup_info.filesystem, backup_info.blob_files)
     backup_info.delete(force_writes=True)
   else:
     kinds_backup_files = tuple(backup_info.get_kind_backup_files())
-    delete_backup_files(backup_info.filesystem, itertools.chain(*(
-        kind_backup_files.files for kind_backup_files in kinds_backup_files)))
+    if delete_files:
+      delete_backup_files(backup_info.filesystem, itertools.chain(*(
+          kind_backup_files.files for kind_backup_files in kinds_backup_files)))
     db.delete(kinds_backup_files + (backup_info,), force_writes=True)
 
 
@@ -616,6 +681,7 @@ class DoBackupRestoreHandler(BaseDoHandler):
   """
   SUFFIX = 'backup_restore.do'
   BACKUP_RESTORE_HANDLER = __name__ + '.RestoreEntity.map'
+  RESTORE_COMPLETE_HANDLER = __name__ + '.RestoreCompleteHandler'
 
   INPUT_READER = input_readers.__name__ + '.RecordsReader'
   _get_html_page = 'do_restore_from_backup.html'
@@ -651,6 +717,7 @@ class DoBackupRestoreHandler(BaseDoHandler):
       mapper_params = self._GetBasicMapperParams()
       mapper_params['files'] = get_backup_files(backup, kinds)
       mapper_params['kind_filter'] = kinds
+      mapper_params['original_app'] = backup.original_app
       mapreduce_params = {
           'backup_name': backup.name,
           'force_ops_writes': True
@@ -666,6 +733,65 @@ class DoBackupRestoreHandler(BaseDoHandler):
         job_operation.status = utils.DatastoreAdminOperation.STATUS_FAILED
         job_operation.put(force_writes=True)
       raise
+    finally:
+
+
+
+      if self.request.get('delete_backup_after_restore', '').lower() == 'true':
+        delete_backup_info(backup, delete_files=False)
+
+
+class DoBackupImportHandler(BaseDoHandler):
+  """Handler to deal with datastore admin requests to import backup info."""
+
+  SUFFIX = 'import_backup.do'
+
+  def get(self):
+    self.post()
+
+  def post(self):
+    """Handler for post requests to datastore_admin/import_backup.do.
+
+    Import is executed and user is redirected to the base-path handler.
+    """
+    gs_handle = self.request.get('gs_handle')
+    token = self.request.get('xsrf_token')
+    error = None
+    if gs_handle and utils.ValidateXsrfToken(token, XSRF_ACTION):
+      try:
+        bucket_name, path = parse_gs_handle(gs_handle)
+        file_content = get_gs_object(bucket_name, path)
+        entities = parse_backup_info_file(file_content)
+        original_backup_info = entities.next()
+        entity = datastore.Entity(BackupInformation.kind())
+        entity.update(original_backup_info)
+        backup_info = BackupInformation.from_entity(entity)
+        if original_backup_info.key().app() != os.getenv('APPLICATION_ID'):
+          backup_info.original_app = original_backup_info.key().app()
+
+        def tx():
+          backup_info.put(force_writes=True)
+          kind_files_models = []
+          for entity in entities:
+            kind_files = backup_info.create_kind_backup_files(
+                entity.key().name(), entity['files'])
+            kind_files_models.append(kind_files)
+          db.put(kind_files_models, force_writes=True)
+        db.run_in_transaction(tx)
+        backup_id = str(backup_info.key())
+      except Exception, e:
+        logging.exception('Failed to Import datastore backup information.')
+        error = str(e)
+
+    if error:
+      query = urllib.urlencode([('error', error)])
+      self.redirect('%s?%s' % (utils.config.BASE_PATH, query))
+    elif self.request.get('Restore'):
+      ConfirmRestoreFromBackupHandler.Render(
+          self, default_backup_id=backup_id,
+          default_delete_backup_after_restore=True)
+    else:
+      self.redirect(utils.config.BASE_PATH)
 
 
 class BackupInformation(db.Model):
@@ -679,6 +805,8 @@ class BackupInformation(db.Model):
   completed_jobs = db.StringListProperty()
   complete_time = db.DateTimeProperty(default=None)
   blob_files = db.StringListProperty()
+  original_app = db.StringProperty(default=None)
+  gs_handle = db.TextProperty(default=None)
 
   @classmethod
   def kind(cls):
@@ -749,17 +877,62 @@ def BackupCompleteHandler(operation, job_id, mapreduce_state):
                                                                  filenames)
       db.put((backup_info, kind_backup_files), force_writes=True)
     if operation.status == utils.DatastoreAdminOperation.STATUS_COMPLETED:
-      finalize_backup_info(backup_info)
+      deferred.defer(finalize_backup_info, backup_info.key(),
+                     mapreduce_spec.mapper.params,
+                     _url=utils.ConfigDefaults.DEFERRED_PATH,
+                     _transactional=True)
   else:
     logging.warn('BackupInfo was not found for %s',
                  mapreduce_spec.params['backup_info_pk'])
 
 
-def finalize_backup_info(backup_info):
-  backup_info.complete_time = datetime.datetime.now()
+def finalize_backup_info(
+    backup_info_pk, mapper_params):
+  """Finalize the state of BackupInformation and creates info file for GS."""
 
-  backup_info.put(force_writes=True)
-  logging.info('Backup %s completed', backup_info.name)
+
+  def tx():
+    backup_info = BackupInformation.get(backup_info_pk)
+    if backup_info:
+      backup_info.complete_time = datetime.datetime.now()
+      if backup_info.filesystem == files.GS_FILESYSTEM:
+
+        key_str = str(backup_info.key()).replace('/', '_')
+        gs_bucket = mapper_params['gs_bucket_name']
+        gs_handle = '/gs/%s/%s.backup_info' % (gs_bucket, key_str)
+        backup_info.gs_handle = gs_handle
+        create_backup_info_file(gs_handle, backup_info)
+      backup_info.put(force_writes=True)
+      logging.info('Backup %s completed', backup_info.name)
+    else:
+      logging.warn('Backup %s could not be found', backup_info_pk)
+  db.run_in_transaction(tx)
+
+
+def create_backup_info_file(filename, backup_info):
+  """Creates a backup_info_file for the given BackupInformation model."""
+  info_file = files.open(files.gs.create(filename), 'a', exclusive_lock=True)
+  try:
+    with records.RecordsWriter(info_file) as writer:
+
+      writer.write('1')
+
+      writer.write(db.model_to_protobuf(backup_info).SerializeToString())
+
+      for kind_files in backup_info.get_kind_backup_files():
+        writer.write(db.model_to_protobuf(kind_files).SerializeToString())
+  finally:
+    info_file.close(finalize=True)
+
+
+def parse_backup_info_file(content):
+  """Returns entities iterator from a backup_info file content."""
+  reader = records.RecordsReader(cStringIO.StringIO(content))
+  version = reader.read()
+  if version != '1':
+    raise IOError('Unsupported version')
+  for record in reader:
+    yield datastore.Entity.FromPb(record)
 
 
 @db.non_transactional
@@ -798,9 +971,20 @@ class RestoreEntity(object):
   """A class which restore the entity to datastore."""
 
   def __init__(self):
+    self.initialized = False
+    self.kind_filter = None
+    self.app_id = None
+
+  def initialize(self):
+    if self.initialized:
+      return
     mapper_params = context.get().mapreduce_spec.mapper.params
     kind_filter = mapper_params.get('kind_filter')
     self.kind_filter = set(kind_filter) if kind_filter else None
+    original_app = mapper_params.get('original_app')
+    if original_app and os.getenv('APPLICATION_ID') != original_app:
+      self.app_id = os.getenv('APPLICATION_ID')
+    self.initialized = True
 
   def map(self, record):
     """Restore entity map handler.
@@ -811,10 +995,15 @@ class RestoreEntity(object):
     Yields:
       A operation.db.Put for the mapped entity
     """
+    self.initialize()
     pb = entity_pb.EntityProto(contents=record)
+    if self.app_id:
+      utils.FixKeys(pb, self.app_id)
     entity = datastore.Entity.FromPb(pb)
     if not self.kind_filter or entity.kind() in self.kind_filter:
       yield op.db.Put(entity)
+      if self.app_id:
+        yield utils.AllocateMaxId(entity.key(), self.app_id)
 
 
 def validate_gs_bucket_name(bucket_name):
@@ -860,6 +1049,51 @@ def is_accessible_bucket_name(bucket_name):
 
 
 
+def parse_gs_handle(gs_handle):
+  """Splits [/gs/]?bucket_name[/folder]*[/file]? to (bucket_name, path | '')."""
+  if gs_handle.startswith('/'):
+    filesystem = gs_handle[1:].split('/', 1)[0]
+    if filesystem == 'gs':
+      gs_handle = gs_handle[4:]
+    else:
+      raise BackupValidationException('Unsupported filesystem: %s' % filesystem)
+  tokens = gs_handle.split('/', 1)
+  return (tokens[0], '') if len(tokens) == 1 else tuple(tokens)
+
+
+def list_bucket_files(
+    bucket_name, prefix, max_keys=1000):
+  """Returns a listing of of a bucket that matches the given prefix."""
+  scope = 'https://www.googleapis.com/auth/devstorage.read_only'
+  url = 'https://%s.commondatastorage.googleapis.com/?' % bucket_name
+  query = [('max-keys', max_keys)]
+  if prefix:
+    query.append(('prefix', prefix))
+  url += urllib.urlencode(query)
+  auth_token, _ = app_identity.get_access_token(scope)
+  result = urlfetch.fetch(url, method=urlfetch.GET, headers={
+      'Authorization': 'OAuth %s' % auth_token,
+      'x-goog-api-version': '2'})
+  if result and result.status_code == 200:
+    doc = xml.dom.minidom.parseString(result.content)
+    return [node.childNodes[0].data for node in doc.getElementsByTagName('Key')]
+  raise BackupValidationException('Request to Google Cloud Storage failed')
+
+
+def get_gs_object(bucket_name, path):
+  """Returns a listing of of a bucket that matches the given prefix."""
+  scope = 'https://www.googleapis.com/auth/devstorage.read_only'
+  url = 'https://%s.commondatastorage.googleapis.com/%s' % (bucket_name, path)
+  auth_token, _ = app_identity.get_access_token(scope)
+  result = urlfetch.fetch(url, method=urlfetch.GET, headers={
+      'Authorization': 'OAuth %s' % auth_token,
+      'x-goog-api-version': '2'})
+  if result and result.status_code == 200:
+    return result.content
+  raise BackupValidationException('Requested path was not found')
+
+
+
 def get_queue_names(app_id=None, max_rows=100):
   """Returns a list with all non-special queue names for app_id."""
   rpc = apiproxy_stub_map.UserRPC('taskqueue')
@@ -896,4 +1130,6 @@ def handlers_list(base_path):
        DoBackupDeleteHandler),
       (r'%s/%s' % (base_path, DoBackupAbortHandler.SUFFIX),
        DoBackupAbortHandler),
+      (r'%s/%s' % (base_path, DoBackupImportHandler.SUFFIX),
+       DoBackupImportHandler),
       ]
